Time: 0 ms (100.00%), Space: 38.9 MB (94.62%) - LeetHub

+++ 509-fibonacci-number/509-fibonacci-number.java
@@ -3,13 +3,13 @@ class Solution {
         if(n==0 || n==1)
             return n;
         
-        long ans[] = new long[n+1];
+        int ans[] = new int[n+1];
         
         ans[0] = 0;
         ans[1] = 1;
         for(int i=2; i<=n; i++){
             ans[i] = ans[i-1] + ans[i-2];
         }
-        return (int)(ans[n]);
+        return ans[n];
     }
 }
